解决两个checkbox无法点击问题

+++ app/src/main/java/com/zs/my_ecommerce/bean/Cart.kt
@@ -8,8 +8,8 @@ data class Cart(
     val id: Int,
     val title: String,
     val description: String,
-    val price: Double,
-    val rating: Double,
+    val price: Float,
+    val rating: Float,
     val thumbnail: String,
     var num: Int = 0,
     @PrimaryKey(autoGenerate = true)
