+++ app/src/main/java/com/zs/my_ecommerce/bean/Favorite.kt
@@ -8,8 +8,8 @@ data class Favorite(
     val id: Int,
     val title: String,
     val description: String,
-    val price: Double,
-    val rating: Double,
+    val price: Float,
+    val rating: Float,
     val thumbnail: String,
     @PrimaryKey(autoGenerate = true)
     val uuid: Int = 0
